tests/requirements.txt
accessible-pygments==0.0.5
alabaster==0.7.16
ansys-sphinx-theme==1.0.11
astroid==3.3.4
asttokens==2.4.1
babel==2.16.0
beautifulsoup4==4.12.3
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
dill==0.3.8
docutils==0.20.1
executing==2.1.0
fonttools==4.54.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.27.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.1.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pydata-sphinx-theme==0.15.4
Pygments==2.18.0
pylint==3.3.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
setuptools==75.1.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==7.4.7
sphinx-autoapi==3.3.2
sphinx-jinja==2.0.2
sphinx-rtd-theme==2.0.0
sphinx_design==0.6.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.16.0
zipp==3.20.2
